+++ src/Entities/Schema.php
@@ -306,6 +306,9 @@ final class Schema implements IteratorAggregate, ArrayAccess
             'previews/webhooks' => function (array $content) {
                 return new Webhook($content);
             },
+            'credential-hashes/webhooks/{hash}' => function (array $content) {
+                return new WebhookCredential($content);
+            },
             'tracking/webhooks' => function (array $content) {
                 return new Collection(new WebhookTracking(), $content);
             },