+++ tests/Api/ServiceTest.php
@@ -932,6 +932,29 @@ class ServiceTest extends BaseTestCase
     /**
      * @test
      */
+    public function webhookCredentialService()
+    {
+        $client = new Client(['apiKey' => 'QWERTY']);
+        $response = $client->createResponse()->withHeader('Location', 'credential-hashes/webhooks/dummy');
+
+        /** @var CurlHandler|MockObject $handler */
+        $handler = $this->createMock(CurlHandler::class);
+
+        $handler->expects($this->any())->method('__invoke')->willReturn($response);
+
+        $client = new Client([
+            'apiKey' => 'QWERTY',
+            'httpHandler' => $handler,
+        ]);
+        $service = $client->webhookCredentials();
+        $webhook = $service->load('dummy');
+
+        self::assertInstanceOf(Entities\WebhookCredential::class, $webhook);
+    }
+
+    /**
+     * @test
+     */
     public function gatewayDowntimeService()
     {
         $client = new Client(['apiKey' => 'QWERTY']);
@@ -1180,6 +1203,11 @@ class ServiceTest extends BaseTestCase
                 'webhooks',
                 Services\WebhooksService::class,
                 Entities\Webhook::class,
+            ],
+            [
+                'webhookCredentials',
+                Services\WebhookCredentialsService::class,
+                Entities\WebhookCredential::class,
             ],
             [
                 'webhooksTracking',